typescript: extract base typings to shared file
